v1_backup/requirements.txt
streamlit
pandas
plotly
requests
python-dotenv
snowflake-connector-python
pyyaml
openai 